requirements.txt
flask
flask-cors
torch
Pillow
